feat(mobile): init SQLite for offline-first support
