+++ apps/mobile/package.json
@@ -27,6 +27,7 @@
     "expo-linking": "~8.0.8",
     "expo-navigation-bar": "^5.0.10",
     "expo-router": "~6.0.10",
+    "expo-sqlite": "~16.0.10",
     "expo-status-bar": "~3.0.8",
     "expo-system-ui": "~6.0.7",
     "expo-web-browser": "~15.0.10",
@@ -34,6 +35,7 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-native": "0.81.5",
+    "react-native-css-interop": "^0.2.1",
     "react-native-gesture-handler": "~2.28.0",
     "react-native-reanimated": "~4.1.1",
     "react-native-safe-area-context": "~5.6.0",
