+++ apps/mobile/src/db/database.ts
@@ -1,0 +1,10 @@
+import * as SQLite from 'expo-sqlite';
+
+let db: SQLite.SQLiteDatabase | null = null;
+
+export async function getDatabase() {
+  if (!db) {
+    db = await SQLite.openDatabaseAsync('blob.db');
+  }
+  return db;
+}
