+++ apps/mobile/src/db/index.ts
@@ -1,0 +1,7 @@
+import { getDatabase } from './database';
+import { runMigrations } from './migrations';
+
+export async function initDatabase() {
+  const db = await getDatabase();
+  await runMigrations(db);
+}
