+++ apps/mobile/src/db/migrations.ts
@@ -1,0 +1,10 @@
+import * as SQLite from 'expo-sqlite';
+
+export async function runMigrations(db: SQLite.SQLiteDatabase) {
+  await db.execAsync(`
+    CREATE TABLE IF NOT EXISTS meta (
+      key TEXT PRIMARY KEY,
+      value TEXT
+    );
+  `);
+}
